extract Viewer agent claim display model
